requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
transformers>=4.36.0
peft>=0.8.2
accelerate>=0.24.1
safetensors>=0.4.1
svgpathtools>=1.6.1
networkx>=3.2.1
Pillow>=10.1.0
numpy>=1.26.2
pandas>=2.1.3
PyYAML>=6.0.1
tqdm>=4.66.1
matplotlib>=3.8.2
opencv-python>=4.8.1.78
shapely>=2.0.2
rich>=13.7.0
openai>=1.12.0
